Don't call a non-test file *test.h.
fips_break_test.h is a bad name because generate_build_scripts.h thinks
that it's a test file, which it is, but one that's needed in the main
build. Thanks to Svilen Kanev for noting this.

That header doesn't particularly carry its weight. The idea was that
rebuilding the break test wouldn't need to rebuild everything if that
logic was isolated in its own header. But we only have to rebuild once
now, so whatever. There's already a block of crypto/internal.h with very
similar stuff; it can go there.

Change-Id: Ifb479eafd4df9a7aac4804cae06ba87257c77fc3