Unexport X509_CERT_AUX and remove X509_CERT_AUX.other
This type is opaque, with no accessors or setters, and there is no way
to get a hold of one except by parsing it. It's only used indirectly via
X509 functions.

The 'other' field is unused and appears to be impossible to set or
query, in either us or upstream.

Change-Id: I4aca665872792f75e9d92e5af68da597b849d4b6