Retire the Windows BIO_printf workaround.
With the UCRT, introduced in VS 2015, vsnprintf in MSVC is now
C99-conformant. See:
https://docs.microsoft.com/en-us/cpp/c-runtime-library/reference/vsnprintf-vsnprintf-vsnprintf-l-vsnwprintf-vsnwprintf-l?view=msvc-170

It is a little unclear to me whether "Beginning with the UCRT in Visual
Studio 2015 and Windows 10" means it is only C99-conformant in Windows
10, or if this is referring to how the UCRT starts becoming an OS
component in Windows 10. I think the latter. This document talks about
the UCRT:
https://docs.microsoft.com/en-us/cpp/porting/upgrade-your-code-to-the-universal-crt?view=msvc-170

But we have tests which cover this in BIOTest.Printf. If it's not
C99-compliant in Windows 7, we'll notice in Chromium's CI.

Change-Id: I932ec2633f94bd77dbe797b06a6bfbc95a568335