Update fiat-crypto.
The files no longer need to be patched because fiat-crypto now has its
own copy of our value barrier. It does, however, require syncing our
NO_ASM define with fiat's. It also trips -Wunused-function in GCC,
because we don't use every function in fiat. I've silenced that with a
pragma.

fiat-crypto is now licensed under any of MIT, BSD 1-clause, or Apache 2.
I've stuck with the MIT one as that's what we were previously importing.

No measurable perf difference before/after this CL, with GCC or Clang on
x86_64.

Change-Id: I2939fd517de37aabdea3ead49150135200a1b112