Tidy up how ASN1_STRING_print_ex figures out the type.
Between the lookup table, the multiple layers of reuse of the "type"
variable, it is a little hard to follow what's going on with
ASN1_STRING_print_ex. Replace the lookup table with a switch-case
(implicitly handles the bounds check, and we can let the compiler figure
out the best spelling). Then, rather than returning a "character width",
which doen't represent UTF-8, just use the already-defined MBSTRING_*
constants.

(These changes should be covered by the existing ASN1Test.StringPrintEx
test.)

Change-Id: Ie3b2557bfae0f65db969e90cd0c76bc8ade963d4