Update build tools.
I'm leaving libc++abi unupdated for now, because it depends on a newer
libc++ than Chromium has managed to update to just yet. (It looks like
Chromium is, for now, manually passing -D_LIBCPP_CONSTINIT=constinit.)
Probably easier to pick that up when the dust settles.

(When that happens, libc++abi will need to be built as C++20, but that's
fine because this is only used on bots.)

Change-Id: I2b7ae62b618b0e09863b4b38548be1b937fbe0d0