Use CMake's C/C++ version features.
This seems to be the preferred mechanism starting CMake 3.1, and reduces
some compiler-specific bits. We can probably also retire -stdlib=libc++
on AppleClang now. https://boringssl-review.googlesource.com/21984
suggests it was needed to make C++11 work, but presumably that's the
default now that the default is itself C++11.

(And if it is needed, that ought to be CMAKE_CXX_STANDARD's
responsibility. I don't see any evidence of them setting it, so I'm
guessing this is just no longer needed.)

Change-Id: Ide1fdbd3d4b8fe37a10a7948171a2c20c9de21f9