Move digestsign.c into the module.
For now, it contains a call to set the service indicator so it should
live in the module. In term it would be nice to move it back out and
have the service indicator set in RSA and ECDSA functions themselves
once the ECDSA functions can take an indicator of the hash function
used.

Change-Id: I2a3c262f66b1881a96ae3e49784a0dc9fc8c4589