rsa: FIPS allows lengths >= 2048 now.
IG C.F now allows RSA modulus sizes >= 2048 to be approved. This change
reflects that in the service indicator and, somewhat, in key generation.
Approved generation is still limited to 4096 bits, however, because
larger values aren't very practical and we would need to increase the
size of the sqrt(2) again.

Change-Id: Ic532b48b8161a1cbc093fb2fa4f6a8fa9b9ac395