Rearrange the C11 vs MSVC check.
Starting version 2019 16.8 (released November 2020), MSVC actually
implements parts of C11, though disabled by default. You have to pass
/std:c11, and then alignas, alignof, and noreturn all work.

When built that way, better to use the real ones, so check for
__STDC_VERSION__ first. It would be nice to mandate that so we can
remove the polyfill, but for now just opportunistically use it when we
can.

Sadly, even in this mode, they still don't implement C11 atomics, so the
refcounting implementation will still be slow.

Change-Id: I28dab4a339c368f7d8f8da5aa7aee1cb344803d3