Fix build with MSVC 2022.
MSVC 2022's C4191 warns on most function pointer casts. Fix and/or
silence them:

connect.c is an unavoidable false positive. We're casting the pointer to
the correct type. The problem was that the caller is required to first
cast it to the wrong type in OpenSSL's API, due to the BIO_callback_ctrl
calling convention. Suppress it.

LHASH_OF(T) and STACK_OF(T)'s macros also now suppress the
warning. Calling the functions through the casted types would indeed
be UB, but we don't do that. We use them as goofy type-erased types,
trying to satisfy a mess of constraints. The problem is there is no
function pointer equivalent of void*.

The sk_sort instance is a true instance of UB. The problem is qsort
lacks a context parameter. I've made sk_sort call qsort_s on _MSC_VER,
to silence the warning.  Ideally we'd fix it on other platforms, but
qsort_r and qsort_s are a disaster. See
https://stackoverflow.com/a/39561369

Change-Id: I0dca80670c74afaa03fc5c8fd7059b4cfadfac72
Fxied: 495