Define STACK_OF(OPENSSL_STRING) more straight-forwardly.
This results in one change, which is the comparison function goes from:

  typedef char *OPENSSL_STRING;
  static int sk_strcmp(const OPENSSL_STRING *a,
                       const OPENSSL_STRING *b);

which is:

  static int sk_strcmp(char *const *a, char *const *b)

into:

  static int sk_strcmp(const char **a, const char **b)

Neither is correct (both consts should be there), but switching the
defintion is necessary to attach the 'const' to 'char' itself. Otherwise
it wouldn't see through the typedef. Fixing the rest of the calling
convention will finish the job.

(While STACK_OF(OPENSSL_STRING) is used externally, nothing external
touches the comparison function.)

Bug: 498
Change-Id: I77bdf2a72b2553bf9409a1d39326890ed5c3582c