Add X509_STORE_CTX_set0_trusted_stack.
OpenSSL renamed X509_STORE_CTX_trusted_stack to
X509_STORE_CTX_set0_trusted_stack. This name is a partially an improvement as
this is a setter, and partially a setback. The "set0" name is a bit
misleading.

This is narrowly correct, in that this function does not adjust
refcounts. But usually set0 functions don't adjust refcounts because
they take ownership of the input. This function does not. It simply
borrows the pointer and assumes it will remain valid for the duration of
X509_STORE_CTX.

Change-Id: I0ef37ba56a2feece6f927f033bdcb4671225dc6f