Add an EVP_HPKE_KEM_enc_len API.
OHTTP concatenates enc to the ciphertext without any length prefix
(unlike ECH), so an implementation would want to know the length of the
KEM. Add an accessor to help with that.

While I'm here, fix a couple places where we assumed a specific KEM in
the HPKE implementation (although we still only support the one KEM so
this is all moot).

There's probably something to be said for lifting the length checks out
of the KEM-specific code and into the wrappers, as we're assuming
fixed-width fields anyway. But I've left it alone for now.

Change-Id: I634a053faa5e3b35d846b690140333bdc741f92a