Require getrandom in all FIPS builds.
It is now 2022. See if we can assume getrandom in this configuration.
For now, I've made a minimal change. If we don't find this needs to be
reverted, we can fully unwind this logic.

Update-Note: The /dev/urandom fallback is no longer available in FIPS
builds. This fallback relied on RNGGETENTCNT and was quite flaky.

Change-Id: Icf6d29f6d5952fb6c5656c9039a4cfaf1de2d724