Silence a GCC 12 -Warray-bounds false positive warning.
GCC 12 triggers a -Warray-bounds false positive in crypto/x509v3's IPv6
parser. Although v6stat.total cannot exceed 16 because of the callback,
GCC doesn't know this and seems to get confused. Checking >= 16 seems to
silence it.

While I'm here, move the comments so they don't obscure the
if/else-if chains and avoid a theoretical overflow in 'zero_cnt' by
checking for the maximum value inside the callback. (Although I'm
confused why this parser thinks IPv6 literals can have so many colons in
a row. I thought two was the maximum. Note 'zero_cnt' counts the number
of zero-length spaces in between the colons, so three would be '::::'.)

Change-Id: If1610a36693915aa92085d8cb3a4709ae82992ba