Reset the key_update_count on writes.
 We reset the key_update_count to 0 any time we write 256 application
data records in between seeing a key update request from our peer. We
contiue as before to also reset the counter to 0 any time we read
application data from the peer.

 This allows for the case where the peer is counting bytes and
requesting key updates when a certain number of bytes have been sent,
but is not actually sending data itself, only reading data.  This is
not a complete fix, because it still depends on the application to poll
and write data, but may avoid issues with peer applications that only
read data and send key updates periodically.

Bug: b:244295650
Bug: 517

Change-Id: Iddb69cf1e8803caf34c82c5deaa069f2bfc12d17