Add the "groups" variants of SSL_CTX_set1_curves_list.
Node calls these. OpenSSL renamed their APIs to align with the IETF
renaming NamedCurve to NamedGroup. (Ironically, with post-quantum
ciphers, that name turns out also to be wrong and it probably should
have been a reference to KEMs.)

To avoid churn for now, I haven't marked the old ones as deprecated, or
renamed any of the internal types yet. We can see about doing that
later.

Change-Id: I5765cea8398f3836611977805bf8ae7d6efc0a70